Moved index.ts to src directory
