--- index.ts
@@ -1,24 +1,0 @@
-export * from "./src/helper-puppeteer";
-
-export {
-  HTMLMarkupChangedError,
-  LoginError,
-  MyTimeoutError,
-  NoInternetError,
-  NotImplementedError,
-} from "./src/custom-errors";
-
-export * from "./src/helper";
-
-export * from "./src/savers/screenshot-saver";
-export * from "./src/savers/screenshot-saver-interface";
-export * from "./src/savers/cookie-saver";
-export * from "./src/savers/cookie-saver-interface";
-
-export * from "./src/puppeteer-utils";
-
-export * from "./src/botbase";
-
-export * from "./src/browser-launcher";
-
-// export * from "./src/utils";
